refactor: extract Header and HeroSection components from CloudIntegrationPage
